requirements.txt
Django==3.1.2
djangorestframework==3.12.1
django-webpack-loader==0.7.0
psycopg2>=2.7,<3.0
uwsgi==2.0.19.1
